+++ ICD.Connect.Misc.CrestronPro/Ports/DigitalInput/DigitalInputAdapterSettings.cs
@@ -14,6 +14,7 @@ namespace ICD.Connect.Misc.CrestronPro.Ports.DigitalInput
 
 		#region Properties
 
+		[ControlPortParentSettingsProperty]
 		[OriginatorIdSettingsProperty(typeof(IPortParent))]
 		public int? Device { get; set; }