+++ ICD.Connect.Misc.GlobalCache/Devices/IP2SL/GcIp2SlDeviceSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Misc.GlobalCache.Devices.IP2SL
+{
+	[KrangSettings("GlobalCacheIP2SL", typeof(GcIp2SlDeviceSettings))]
+	public sealed class GcIp2SlDeviceSettings : AbstractGcITachDeviceSettings
+	{
+	}
+}