+++ ICD.Connect.Misc.GlobalCache/Ports/IGcITachPort.cs
@@ -1,0 +1,8 @@
+﻿using ICD.Connect.Protocol.Ports;
+
+namespace ICD.Connect.Misc.GlobalCache.Ports
+{
+	public interface IGcITachPort : IPort
+	{
+	}
+}